terms&conditions added file allowed extensions add

+++ client/src/components/Auth/GoogleAuth.tsx
@@ -13,11 +13,12 @@ interface User {
   email: string;
   avatar: string;
 }
+interface GoogleProps {
+  isTermsConfirmed: boolean;
+}
 
-function GoogleAuth() {
+function GoogleAuth({ isTermsConfirmed }: GoogleProps) {
   const [user, setUser] = useState<User | null>(null);
-  // const { fetchData, responseData } = useFetchApi<ApiResponse>();
-
   // useEffect(() => {
   //   if (responseData?.user) {
   //     localStorage.setItem("user", JSON.stringify(responseData?.user));
@@ -27,13 +28,20 @@ function GoogleAuth() {
   // }, [responseData]);
 
   const onGoogleSuccess = (credentialResponse: any) => {
-    callApi("/api/sessions/login", "POST", {
-      credential: credentialResponse.credential,
-    }).then((data) => {
-      localStorage.setItem("user", JSON.stringify(data?.user));
-      setUser(data.user);
-      window.location.reload();
-    });
+    console.log("isconfirmed", isTermsConfirmed);
+    if (isTermsConfirmed) {
+      callApi("/api/sessions/login", "POST", {
+        credential: credentialResponse.credential,
+      }).then((data) => {
+        localStorage.setItem("user", JSON.stringify(data?.user));
+        setUser(data.user);
+        window.location.reload();
+      });
+    } else {
+      alert(
+        "Please read the terms and conditions and click 'agree' before signing in."
+      );
+    }
   };
 
   return (
